sdp4_message_bot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
frozenlist==1.7.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.6.4
propcache==0.3.2
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
sniffio==1.3.1
typing_extensions==4.13.0
yarl==1.20.1
